Moved classes
Moved classes to their own folder, and importing them in main.py.
