+++ main.py
@@ -1,143 +1,9 @@
 # Example file showing a circle moving on screen
 import pygame as pg
-import pygame.gfxdraw as gfx
+from classes.curves import Curve, Curve_point
+from classes.path import Path
+from classes.vector import Vector_layer
 
-
-# A single point on a curve
-class Curve_point:
-    def __init__(self, position, is_control=False):
-        position = pg.math.Vector2(position)
-        self.x = position.x
-        self.y = position.y
-        self.is_control = is_control
-        self.collider = pg.Rect(self.x - 5, self.y - 5, 10, 10)
-
-    def get_position(self):
-        return pg.Vector2(self.x, self.y)
-
-
-# A single curve segment
-class Curve:
-    def __init__(self, start, first_control, second_control, end, steps=100, color=(0, 0, 0)):
-        first_control.is_control = True
-        second_control.is_control = True
-        self.points = [start, first_control, second_control, end]
-        self.steps = steps
-        self.color = color
-
-    def draw(self, surface, position):
-        gfx.bezier(surface, [(p.x + position.x, p.y + position.y)
-                   for p in self.points], self.steps, self.color)
-
-    def draw_controls(self, surface, offset_position):
-        pg.draw.line(surface, "darkgray", (self.points[0].x + offset_position.x, self.points[0].y +
-                     offset_position.y), (self.points[1].x + offset_position.x, self.points[1].y + offset_position.y))
-        pg.draw.line(surface, "darkgray", (self.points[2].x + offset_position.x, self.points[2].y +
-                     offset_position.y), (self.points[3].x + offset_position.x, self.points[3].y + offset_position.y))
-        for point in self.points:
-            color = "cadetblue" if point.is_control else "black"
-            pg.draw.circle(surface, color, (point.x +
-                           offset_position.x, point.y + offset_position.y), 5)
-
-    def update_point_colliders(self, offset_position):
-        for point in self.points:
-            point.collider.center = (
-                point.x + offset_position.x, point.y + offset_position.y)
-
-    def bounding_box(self):
-        x_min = min([p.x for p in self.points])
-        x_max = max([p.x for p in self.points])
-        y_min = min([p.y for p in self.points])
-        y_max = max([p.y for p in self.points])
-        return pg.Rect(x_min, y_min, x_max - x_min, y_max - y_min)
-
-    def get_points(self):
-        return self.points
-    
-    def start_point(self):
-        return self.points[0]
-    
-    def end_point(self):
-        return self.points[3]
-
-# A composite of multiple curves
-
-
-class Path:
-    def __init__(self, start_segment=None, position=(0, 0)):
-        self.segments = [start_segment] if not start_segment is None else []
-        self.position = pg.Vector2(position)
-        self.first_point = start_segment.start_point() if not start_segment is None else None
-        self.last_point = start_segment.end_point() if not start_segment is None else None
-
-    def add_segment_to_point(self, point):
-        if self.last_point is None:
-            self.first_point = point
-            self.last_point = point
-        else:
-            distance = self.last_point.get_position().distance_to(point.get_position())
-            start = self.last_point
-            first_control = Curve_point(start.get_position().move_towards(point.get_position(), distance/3))
-            second_control = Curve_point(point.get_position().move_towards(start.get_position(), distance/3))
-            self.segments.append(Curve(start, first_control, second_control, point))
-            self.last_point = point
-
-    def close_path(self):
-        if not self.first_point is None and not self.last_point is None:
-            distance = self.last_point.get_position().distance_to(self.first_point.get_position())
-            start = self.last_point
-            first_control = Curve_point(start.get_position().move_towards(self.first_point.get_position(), distance/3))
-            second_control = Curve_point(self.first_point.get_position().move_towards(start.get_position(), distance/3))
-            self.segments.append(Curve(start, first_control, second_control, self.first_point))
-            self.last_point = self.first_point
-
-    def draw(self, surface):
-        for segment in self.segments:
-            segment.draw(surface, self.position)
-
-    def draw_controls(self, surface):
-        for segment in self.segments:
-            segment.draw_controls(surface, self.position)
-
-    def update_point_colliders(self):
-        for segment in self.segments:
-            segment.update_point_colliders(self.position)
-
-    def get_points(self):
-        points = []
-        for segment in self.segments:
-            for point in segment.points:
-                points.append(point)
-        return points
-
-    def large_bounding_box(self):
-        bounding_box = pg.Rect(0, 0, 0, 0)
-        for segment in self.segments:
-            bounding_box.union_ip(segment.bounding_box())
-        return bounding_box.scale_by(1.1)
-    
-
-
-# A layer of vector objects
-class Vector_layer:
-    def __init__(self, id, depth, viewport, colorkey=(0, 255, 255)):
-        self.id = id
-        self.depth = depth
-        self.width = viewport.get_width()
-        self.height = viewport.get_height()
-        self.surface = pg.Surface((self.width, self.height))
-        self.surface.set_colorkey(colorkey)
-        self.objects = []
-
-    def add_object(self, obj):
-        self.objects.append(obj)
-
-    def draw(self):
-        for obj in self.objects:
-            obj.draw(self.surface)
-
-    def clear(self):
-        self.surface.fill((0, 255, 255))
 
 def main():
     # pyg setup
